requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.3.1
constantly==15.1.0
cryptography==41.0.4
daphne==3.0.2
Django==4.2.6
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Pillow==10.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.7
zope.interface==6.1
django-cors-headers
